pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
[project]
name = "tinq"
version = "0.0.2"
authors = [
  { name="Boulama K.", email="boulama@tinq.ai" },
]
description = "A Python wrapper for the Tinq.ai API - an easy-to-use text analysis and natural language processing toolkit."
readme = "README.md"
license = { file="LICENSE" }
requires-python = ">=3.7"
dependencies = ["python-dotenv", "requests"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.urls]
"Homepage" = "https://github.com/tinq-ai/tinq-python"
"Bug Tracker" = "https://github.com/tinq-ai/tinq-python/issues"